+++ lib/g3-types/src/net/tcp/sockopt.rs
@@ -16,9 +16,9 @@ use g3_std_ext::core::OptionExt;
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct TcpMiscSockOpts {
     pub no_delay: Option<bool>,
-    pub max_segment_size: Option<u32>,
-    pub time_to_live: Option<u32>,
-    pub hop_limit: Option<u32>,
+    pub max_segment_size: Option<u16>,
+    pub time_to_live: Option<u8>,
+    pub hop_limit: Option<u8>,
     pub type_of_service: Option<u8>,
     #[cfg(not(windows))]
     pub traffic_class: Option<u8>,
